fix build error

+++ ftl/old/ftl_command.hh
@@ -33,7 +33,7 @@ typedef struct _Command {
   Addr ppn;
   PAL_OPERATION operation;
   bool mergeSnapshot;
-  size_t size;
+  uint64_t size;
 
   _Command()
       : arrived(0),
@@ -42,7 +42,7 @@ typedef struct _Command {
         operation(OPER_NUM),
         mergeSnapshot(false),
         size(0) {}
-  _Command(Tick t, Addr a, PAL_OPERATION op, size_t s)
+  _Command(Tick t, Addr a, PAL_OPERATION op, uint64_t s)
       : arrived(t),
         finished(0),
         ppn(a),
